Restructure module dependencies of v1/v2 adaround
