+++ TrainingExtensions/torch/src/python/aimet_torch/_base/seq_mse.py
@@ -53,7 +53,7 @@ from aimet_common.utils import AimetLogger
 
 from aimet_torch.utils import CachedDataset, get_ordered_list_of_modules, in_eval_mode, StopForwardException,\
     change_tensor_device_placement, get_device
-from aimet_torch.v1.adaround.activation_sampler import create_modulelist_for_group_modules,\
+from aimet_torch._base.adaround.activation_sampler import create_modulelist_for_group_modules,\
     get_block_inputs, get_block_outputs
 
 # The following modules with weights are supported
