+++ TrainingExtensions/torch/src/python/aimet_torch/v1/adaround/adaround_wrapper.py
@@ -37,7 +37,6 @@
 
 """ Custom Wrapper for quantizing weights using Adaround """
 
-import abc
 import contextlib
 from typing import Tuple
 import torch
@@ -45,73 +44,16 @@ import torch.nn
 
 # Import AIMET specific modules
 import aimet_common.aimet_tensor_quantizer as AimetTensorQuantizer
-from aimet_common.defs import AdaroundConstants, MAP_QUANT_SCHEME_TO_PYMO
+from aimet_common.defs import MAP_QUANT_SCHEME_TO_PYMO
+from aimet_torch._base.adaround.adaround_wrapper import AdaroundWrapperBase
 from aimet_torch.v1.tensor_quantizer import StaticGridPerChannelQuantizer
-from aimet_torch.v1.qc_quantize_op import QcQuantizeWrapper
 from aimet_torch.v1.quantsim_straight_through_grad import broadcast_to_tensor
-from aimet_torch.v2.utils import patch_attr
 
-
-class AdaroundWrapperBase(abc.ABC, torch.nn.Module):
-    """
-    Adaround base class
-    """
-    @abc.abstractmethod
-    def forward(self, *args, **kwargs):
-        """
-        Apply adaround and run forward function of the wrapped module
-        """
-
-    @abc.abstractmethod
-    def get_original_module(self) -> torch.nn.Module:
-        """
-        Returns original module so that we can check its
-        module type or access its weight
-        """
-
-    @abc.abstractmethod
-    def apply_adaround(self, tensor: torch.Tensor) -> torch.Tensor:
-        """
-        Apply adaround to the input tensor
-        """
-
-    @property
-    def weight(self) -> torch.Tensor:
-        """
-        Returns the weight of the original model
-        """
-        return getattr(self.get_original_module(), self.weight_name)
-
-    @property
-    def weight_name(self) -> str:
-        """
-        Returns the name of the weight to apply adaround
-        """
-        return 'weight'
 
 class AdaroundWrapper(AdaroundWrapperBase):
     """
-    Basic adaround wrapper class for AIMET v1
+    Adaround wrapper class for AIMET v1
     """
-    def __init__(self, module: QcQuantizeWrapper):
-        super().__init__()
-        assert self.weight_name in module.param_quantizers
-        self.module_to_wrap = module
-        self._init_param()
-
-    def forward(self, *args, **kwargs):
-        """
-        Temporarily replace weight of the wrapped module by adarounded weight
-        and run forward function of wrapped module
-        """
-        origianl_module = self.get_original_module()
-        weight = self.weight
-        if self._is_weight_quantizer_enabled():
-            weight = self.apply_adaround(weight)
-
-        with self._disable_weight_quantizer(), \
-            patch_attr(origianl_module, self.weight_name, weight):
-            return self.module_to_wrap.forward(*args, **kwargs)
 
     def get_original_module(self) -> torch.nn.Module:
         """
@@ -120,33 +62,6 @@ class AdaroundWrapper(AdaroundWrapperBase):
         """
         # pylint: disable=protected-access
         return self.module_to_wrap._module_to_wrap
-
-    def apply_adaround(self, tensor: torch.Tensor) -> torch.Tensor:
-        """
-        Apply adaround to the input tensor
-        """
-        input_dtype = tensor.dtype
-        alpha = self.alpha.to(device=tensor.device, dtype=tensor.dtype)
-
-        # Scale the tensor
-        tensor = torch.floor(tensor / self.broadcasted_delta)
-
-        # Soft rounding maps alpha parameter between zero and one using
-        # rectified sigmoid function and hard rounding maps it to exactly zero or one
-        if self.use_soft_rounding:
-            h_alpha = torch.clamp(torch.sigmoid(alpha) * (AdaroundConstants.ZETA - AdaroundConstants.GAMMA) +
-                                  AdaroundConstants.GAMMA, 0, 1)
-        else:
-            h_alpha = (alpha >= 0).to(tensor.dtype)
-
-        # Adaround the tensor
-        tensor = tensor + h_alpha
-
-        # Quantize and de-quantize the tensor
-        tensor_quant = torch.clamp(tensor - self.broadcasted_offset, self.clip_min, self.clip_max)
-        tensor_dequant = (tensor_quant + self.broadcasted_offset) * self.broadcasted_delta
-
-        return tensor_dequant.to(input_dtype)
 
     @contextlib.contextmanager
     def _disable_weight_quantizer(self):
@@ -190,35 +105,3 @@ class AdaroundWrapper(AdaroundWrapperBase):
 
         ch_axis = self._get_weight_quantizer_channel_axis()
         return broadcast_to_tensor(self.weight, delta, ch_axis), broadcast_to_tensor(self.weight, offset, ch_axis)
-
-    def _get_weight_quantizer_bitwidth(self) -> int:
-        """
-        Returns bitwidth of the weight quantizer
-        """
-        quantizer = self.module_to_wrap.param_quantizers[self.weight_name]
-        return quantizer.bitwidth
-
-    def _init_param(self):
-        """
-        Initialize adaround parameter using the original module
-        """
-        self.broadcasted_delta, self.broadcasted_offset = self._get_weight_quantizer_delta_and_offset()
-        self.alpha = self._generate_alpha_parameter(self.weight, self.broadcasted_delta)
-        self.bitwidth = self._get_weight_quantizer_bitwidth()
-        self.use_soft_rounding = True
-        self.clip_max = 2 ** self.bitwidth - 1
-        self.clip_min = 0
-
-    @staticmethod
-    def _generate_alpha_parameter(tensor: torch.Tensor, delta: torch.Tensor) -> torch.nn.Parameter:
-        """
-        Initializes alpha parameter, same shape as the weight tensor
-        :param tensor: The weight tensor to be ada rounded
-        """
-        tensor_floor = torch.floor(tensor / delta)
-        tensor = (tensor / delta) - tensor_floor
-        alpha = - torch.log((AdaroundConstants.ZETA - AdaroundConstants.GAMMA) / (tensor - AdaroundConstants.GAMMA) - 1)
-
-        # Even if the input is float16, alpha has to be kept in float32
-        # in order to be updated by the optimizer
-        return torch.nn.Parameter(alpha.float(), requires_grad=True)
