+++ TrainingExtensions/torch/test/python/v2/test_adaround.py
@@ -54,7 +54,8 @@ from aimet_common.quantsim import calculate_delta_offset
 from aimet_torch.v1.adaround.adaround_wrapper import AdaroundWrapper
 from aimet_torch.utils import create_fake_data_loader, create_rand_tensors_given_shapes, get_device
 from .models_ import test_models
-from aimet_torch.v1.adaround.adaround_weight import AdaroundOptimizer, AdaroundParameters
+from aimet_torch._base.adaround.adaround_optimizer import AdaroundOptimizer
+from aimet_torch.v1.adaround.adaround_weight import AdaroundParameters
 from aimet_torch.v2.quantsim import QuantizationSimModel
 from aimet_torch.v2.adaround import Adaround
 from aimet_torch.v2.nn import BaseQuantizationMixin
